@@ -6,42 +6,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.12",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "^7.2.15",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "~7.2.0",
+    "@angular/forms": "~8.2.14",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
     "@ng-bootstrap/ng-bootstrap": "^4.1.1",
     "bootstrap": "^4.3.1",
     "core-js": "^2.5.4",
     "file-saver": "^2.0.2",
     "hammerjs": "^2.0.8",
     "jquery": "^3.4.0",
-    "material-community-components": "^3.4.0",
-    "mxgraph": "^3.9.12",
+    "material-community-components": "^3.4.3",
+    "mxgraph": "^4.2.0",
     "ng-sidebar": "^8.0.0",
     "ngx-bootstrap": "^4.0.0",
     "popper.js": "^1.15.0",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.5",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.7",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.803.28",
+    "@angular/cli": "~8.3.28",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
     "@types/file-saver": "^2.0.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "^2.0.7",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -53,7 +53,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2",
+    "typescript": "~3.5.3",
     "webpack-bundle-analyzer": "^3.6.0"
   }
 }
